+++ packages/xchain-ethereum/__tests__/client.testnet.test.ts
@@ -563,6 +563,7 @@ describe('Client Test', () => {
     })
 
     beforeEach(() => {
+      nock.disableNetConnect()
       mock_all_api(etherscanUrl, ropstenInfuraUrl, ropstenAlchemyUrl, 'eth_blockNumber', '0xa7cac8')
       mock_all_api(etherscanUrl, ropstenInfuraUrl, ropstenAlchemyUrl, 'eth_getTransactionCount', '0x0')
       mock_all_api(etherscanUrl, ropstenInfuraUrl, ropstenAlchemyUrl, 'eth_gasPrice', '0x5969ec91')
@@ -611,8 +612,10 @@ describe('Client Test', () => {
     mock_all_api(etherscanUrl, ropstenInfuraUrl, ropstenAlchemyUrl, 'eth_gasPrice', '0xb2d05e00')
     mock_all_api(etherscanUrl, ropstenInfuraUrl, ropstenAlchemyUrl, 'eth_estimateGas', '0x5208')
 
+    const fromAddress = ethClient.getAddress(0)
+
     const gasLimit = await ethClient.estimateApprove({
-      walletIndex: 0,
+      fromAddress,
       contractAddress: '0x8c2a90d36ec9f745c9b28b588cba5e2a978a1656',
       spenderAddress: '0xdac17f958d2ee523a2206206994597c13d831ec7',
       amount: baseAmount(100, ETH_DECIMAL),